Break the implementation into modules
